Changed the table tracks to have DOUBLE type for duration_second. Did the same change for Track.cs. In UserProfileResponse.cs changed the date type to DateTimeOffset for better support of multiple time zones.

+++ Core/Models/Track.cs
@@ -6,5 +6,5 @@ public class Track
     public string Title { get; set; } = string.Empty;
     public string Artist { get; set; } = string.Empty;
     public string FileName { get; set; } = string.Empty;
-    public int DurationSeconds { get; set; }
+    public double DurationSeconds { get; set; }
 }
